requirements.txt
python-telegram-bot==21.10
gunicorn==21.2.0
python-dotenv==1.0.1
